WL-135 added the migration for the expiration_date field in the coupons model
added the functionality of adding the expiration date, and also handled in the shopping cart when the user try to add the expired coupon code in the shopping cart

added expiration date in the coupons list in the ecommerce page

added the unit tests and jasmine tests

added the api to create the course mode

bok-choy tests.

added the test cases for course_modes

added the form for the course modes

fix the quality issues and rebased with master

Renamed "add_honor_mode_to_course" to "add_mode_to_course" since we've removed the hard-coding.
Made the last "/" optional to avoid unnecessary redirection.

regenerated the migration file with a different number

i18n and verified te course mode